# contracts/openapi/planner-v1.yaml
openapi: 3.1.0
info:
  title: 토도록 Planner API
  version: 1.0.0
servers:
  - url: /api/planner/v1
tags:
  - name: Auth
  - name: Calendar
  - name: Task
paths:
  /auth/login:
    post:
      tags: [Auth]
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200': { $ref: '#/components/responses/SessionSuccess' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/refresh:
    post:
      tags: [Auth]
      operationId: refreshSession
      responses:
        '200': { $ref: '#/components/responses/SessionSuccess' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/logout:
    post:
      tags: [Auth]
      operationId: logout
      responses:
        '204': { description: 로그아웃 성공 }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /calendar:
    get:
      tags: [Calendar]
      operationId: getCalendarSummary
      description: 시작일과 종료일을 포함한 최대 42일의 달력 요약을 반환한다.
      parameters:
        - { $ref: '#/components/parameters/FromDate' }
        - { $ref: '#/components/parameters/ToDate' }
      responses:
        '200':
          description: 달력 요약
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CalendarSummaryResponse' }
        '400':
          description: 날짜 범위가 42일을 초과하거나 잘못된 요청
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/ProblemDetails' }
  /calendar/{date}:
    get:
      tags: [Calendar]
      operationId: getDayDetail
      parameters:
        - { $ref: '#/components/parameters/DatePath' }
      responses:
        '200':
          description: 선택 날짜 상세
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DayDetailResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
  /tasks:
    post:
      tags: [Task]
      operationId: createTask
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateTaskRequest' }
      responses:
        '201':
          description: Task 생성
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TaskResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /tasks/{taskId}:
    patch:
      tags: [Task]
      operationId: updateTask
      parameters:
        - { $ref: '#/components/parameters/TaskIdPath' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateTaskRequest' }
      responses:
        '200':
          description: Task 수정
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TaskResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
    delete:
      tags: [Task]
      operationId: deleteTask
      parameters:
        - { $ref: '#/components/parameters/TaskIdPath' }
        - { $ref: '#/components/parameters/VersionQuery' }
      responses:
        '204': { description: Task 삭제 }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /tasks/{taskId}/complete:
    post:
      tags: [Task]
      operationId: completeTask
      parameters:
        - { $ref: '#/components/parameters/TaskIdPath' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/VersionCommand' }
      responses:
        '200':
          description: Task 완료
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TaskResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /tasks/{taskId}/reopen:
    post:
      tags: [Task]
      operationId: reopenTask
      parameters:
        - { $ref: '#/components/parameters/TaskIdPath' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/VersionCommand' }
      responses:
        '200':
          description: Task 재개
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TaskResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
components:
  parameters:
    FromDate:
      name: from
      in: query
      required: true
      schema: { type: string, format: date }
    ToDate:
      name: to
      in: query
      required: true
      description: from부터 최대 42일 범위의 종료일
      schema: { type: string, format: date }
    DatePath:
      name: date
      in: path
      required: true
      schema: { type: string, format: date }
    TaskIdPath:
      name: taskId
      in: path
      required: true
      schema: { type: string, format: uuid }
    VersionQuery:
      name: version
      in: query
      required: true
      schema: { type: integer, format: int64, minimum: 0 }
  responses:
    SessionSuccess:
      description: 세션 발급 성공
      content:
        application/json:
          schema: { $ref: '#/components/schemas/SessionResponse' }
    BadRequest:
      description: 잘못된 요청
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
    Unauthorized:
      description: 인증 실패
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
    NotFound:
      description: 대상을 찾을 수 없음
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
    Conflict:
      description: 버전 충돌
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
  schemas:
    ProblemDetails:
      $ref: './common-v1.yaml#/components/schemas/ProblemDetails'
    LoginRequest:
      type: object
      additionalProperties: false
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
    SessionResponse:
      type: object
      additionalProperties: false
      required: [accessToken, expiresAt, userId]
      properties:
        accessToken: { type: string }
        expiresAt: { type: string, format: date-time }
        userId: { type: string, format: uuid }
    TaskType:
      type: string
      enum: [TODO, WORKOUT, STUDY, CLIMBING]
    TaskStatus:
      type: string
      enum: [PLANNED, COMPLETED, SKIPPED, DELETED]
    TaskResponse:
      type: object
      additionalProperties: false
      required: [taskId, userId, title, taskType, scheduledDate, status, version]
      properties:
        taskId: { type: string, format: uuid }
        userId: { type: string, format: uuid }
        title: { type: string }
        taskType: { $ref: '#/components/schemas/TaskType' }
        scheduledDate: { type: string, format: date }
        status: { $ref: '#/components/schemas/TaskStatus' }
        version: { type: integer, format: int64, minimum: 0 }
    CreateTaskRequest:
      type: object
      additionalProperties: false
      required: [title, taskType, scheduledDate]
      properties:
        title: { type: string, minLength: 1, maxLength: 120 }
        taskType: { $ref: '#/components/schemas/TaskType' }
        scheduledDate: { type: string, format: date }
    UpdateTaskRequest:
      type: object
      additionalProperties: false
      required: [title, scheduledDate, version]
      properties:
        title: { type: string, minLength: 1, maxLength: 120 }
        scheduledDate: { type: string, format: date }
        version: { type: integer, format: int64, minimum: 0 }
    VersionCommand:
      type: object
      additionalProperties: false
      required: [version]
      properties:
        version: { type: integer, format: int64, minimum: 0 }
    CalendarDaySummary:
      type: object
      additionalProperties: false
      required: [date, totalCount, completedCount]
      properties:
        date: { type: string, format: date }
        totalCount: { type: integer, minimum: 0 }
        completedCount: { type: integer, minimum: 0 }
    CalendarSummaryResponse:
      type: object
      additionalProperties: false
      required: [from, to, days]
      properties:
        from: { type: string, format: date }
        to: { type: string, format: date }
        days:
          type: array
          maxItems: 42
          items: { $ref: '#/components/schemas/CalendarDaySummary' }
    DayDetailResponse:
      type: object
      additionalProperties: false
      required: [date, tasks]
      properties:
        date: { type: string, format: date }
        tasks:
          type: array
          items: { $ref: '#/components/schemas/TaskResponse' }