# packages/api-client/src/generated/planner/src/models/CalendarDaySummary.ts
/* tslint:disable */
/* eslint-disable */
/**
 * 토도록 Planner API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface CalendarDaySummary
 */
export interface CalendarDaySummary {
    /**
     * 
     * @type {Date}
     * @memberof CalendarDaySummary
     */
    date: Date;
    /**
     * 
     * @type {number}
     * @memberof CalendarDaySummary
     */
    totalCount: number;
    /**
     * 
     * @type {number}
     * @memberof CalendarDaySummary
     */
    completedCount: number;
}

/**
 * Check if a given object implements the CalendarDaySummary interface.
 */
export function instanceOfCalendarDaySummary(value: object): value is CalendarDaySummary {
    if (!('date' in value) || value['date'] === undefined) return false;
    if (!('totalCount' in value) || value['totalCount'] === undefined) return false;
    if (!('completedCount' in value) || value['completedCount'] === undefined) return false;
    return true;
}

export function CalendarDaySummaryFromJSON(json: any): CalendarDaySummary {
    return CalendarDaySummaryFromJSONTyped(json, false);
}

export function CalendarDaySummaryFromJSONTyped(json: any, ignoreDiscriminator: boolean): CalendarDaySummary {
    if (json == null) {
        return json;
    }
    return {
        
        'date': (new Date(json['date'])),
        'totalCount': json['totalCount'],
        'completedCount': json['completedCount'],
    };
}

export function CalendarDaySummaryToJSON(json: any): CalendarDaySummary {
    return CalendarDaySummaryToJSONTyped(json, false);
}

export function CalendarDaySummaryToJSONTyped(value?: CalendarDaySummary | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'date': value['date'].toISOString().substring(0,10),
        'totalCount': value['totalCount'],
        'completedCount': value['completedCount'],
    };
}