# packages/api-client/src/generated/planner/src/models/SessionResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * 토도록 Planner API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface SessionResponse
 */
export interface SessionResponse {
    /**
     * 
     * @type {string}
     * @memberof SessionResponse
     */
    accessToken: string;
    /**
     * 
     * @type {Date}
     * @memberof SessionResponse
     */
    expiresAt: Date;
    /**
     * 
     * @type {string}
     * @memberof SessionResponse
     */
    userId: string;
}

/**
 * Check if a given object implements the SessionResponse interface.
 */
export function instanceOfSessionResponse(value: object): value is SessionResponse {
    if (!('accessToken' in value) || value['accessToken'] === undefined) return false;
    if (!('expiresAt' in value) || value['expiresAt'] === undefined) return false;
    if (!('userId' in value) || value['userId'] === undefined) return false;
    return true;
}

export function SessionResponseFromJSON(json: any): SessionResponse {
    return SessionResponseFromJSONTyped(json, false);
}

export function SessionResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): SessionResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'accessToken': json['accessToken'],
        'expiresAt': (new Date(json['expiresAt'])),
        'userId': json['userId'],
    };
}

export function SessionResponseToJSON(json: any): SessionResponse {
    return SessionResponseToJSONTyped(json, false);
}

export function SessionResponseToJSONTyped(value?: SessionResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'accessToken': value['accessToken'],
        'expiresAt': value['expiresAt'].toISOString(),
        'userId': value['userId'],
    };
}