# services/planner-service/src/generated/java/io/todorok/planner/api/CalendarApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.24.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package io.todorok.planner.api;

import io.todorok.planner.api.model.CalendarSummaryResponse;
import org.springframework.format.annotation.DateTimeFormat;
import io.todorok.planner.api.model.DayDetailResponse;
import java.time.LocalDate;
import io.todorok.planner.api.model.ProblemDetails;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", comments = "Generator version: 7.24.0")
@Validated
public interface CalendarApi {

    String PATH_GET_CALENDAR_SUMMARY = "/calendar";
    /**
     * GET /calendar
     * 시작일과 종료일을 포함한 최대 42일의 달력 요약을 반환한다.
     *
     * @param from  (required)
     * @param to from부터 최대 42일 범위의 종료일 (required)
     * @return 달력 요약 (status code 200)
     *         or 날짜 범위가 42일을 초과하거나 잘못된 요청 (status code 400)
     */
    @RequestMapping(
        method = RequestMethod.GET,
        value = CalendarApi.PATH_GET_CALENDAR_SUMMARY,
        produces = { "application/json", "application/problem+json" }
    )
    ResponseEntity<CalendarSummaryResponse> getCalendarSummary(
        @NotNull  @Valid @RequestParam(value = "from", required = true) @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) LocalDate from,
        @NotNull  @Valid @RequestParam(value = "to", required = true) @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) LocalDate to
    );


    String PATH_GET_DAY_DETAIL = "/calendar/{date}";
    /**
     * GET /calendar/{date}
     *
     * @param date  (required)
     * @return 선택 날짜 상세 (status code 200)
     *         or 대상을 찾을 수 없음 (status code 404)
     */
    @RequestMapping(
        method = RequestMethod.GET,
        value = CalendarApi.PATH_GET_DAY_DETAIL,
        produces = { "application/json", "application/problem+json" }
    )
    ResponseEntity<DayDetailResponse> getDayDetail(
         @PathVariable("date") @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) LocalDate date
    );

}