# services/planner-service/src/generated/java/io/todorok/planner/api/TaskApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.24.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package io.todorok.planner.api;

import io.todorok.planner.api.model.CreateTaskRequest;
import io.todorok.planner.api.model.ProblemDetails;
import io.todorok.planner.api.model.TaskResponse;
import java.util.UUID;
import io.todorok.planner.api.model.UpdateTaskRequest;
import io.todorok.planner.api.model.VersionCommand;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", comments = "Generator version: 7.24.0")
@Validated
public interface TaskApi {

    String PATH_COMPLETE_TASK = "/tasks/{taskId}/complete";
    /**
     * POST /tasks/{taskId}/complete
     *
     * @param taskId  (required)
     * @param versionCommand  (required)
     * @return Task 완료 (status code 200)
     *         or 대상을 찾을 수 없음 (status code 404)
     *         or 버전 충돌 (status code 409)
     */
    @RequestMapping(
        method = RequestMethod.POST,
        value = TaskApi.PATH_COMPLETE_TASK,
        produces = { "application/json", "application/problem+json" },
        consumes = { "application/json" }
    )
    ResponseEntity<TaskResponse> completeTask(
         @PathVariable("taskId") UUID taskId,
         @Valid @RequestBody VersionCommand versionCommand
    );


    String PATH_CREATE_TASK = "/tasks";
    /**
     * POST /tasks
     *
     * @param createTaskRequest  (required)
     * @return Task 생성 (status code 201)
     *         or 잘못된 요청 (status code 400)
     */
    @RequestMapping(
        method = RequestMethod.POST,
        value = TaskApi.PATH_CREATE_TASK,
        produces = { "application/json", "application/problem+json" },
        consumes = { "application/json" }
    )
    ResponseEntity<TaskResponse> createTask(
         @Valid @RequestBody CreateTaskRequest createTaskRequest
    );


    String PATH_DELETE_TASK = "/tasks/{taskId}";
    /**
     * DELETE /tasks/{taskId}
     *
     * @param taskId  (required)
     * @param version  (required)
     * @return Task 삭제 (status code 204)
     *         or 대상을 찾을 수 없음 (status code 404)
     *         or 버전 충돌 (status code 409)
     */
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = TaskApi.PATH_DELETE_TASK,
        produces = { "application/problem+json" }
    )
    ResponseEntity<Void> deleteTask(
         @PathVariable("taskId") UUID taskId,
        @NotNull @Min(value = 0L)  @Valid @RequestParam(value = "version", required = true) Long version
    );


    String PATH_REOPEN_TASK = "/tasks/{taskId}/reopen";
    /**
     * POST /tasks/{taskId}/reopen
     *
     * @param taskId  (required)
     * @param versionCommand  (required)
     * @return Task 재개 (status code 200)
     *         or 대상을 찾을 수 없음 (status code 404)
     *         or 버전 충돌 (status code 409)
     */
    @RequestMapping(
        method = RequestMethod.POST,
        value = TaskApi.PATH_REOPEN_TASK,
        produces = { "application/json", "application/problem+json" },
        consumes = { "application/json" }
    )
    ResponseEntity<TaskResponse> reopenTask(
         @PathVariable("taskId") UUID taskId,
         @Valid @RequestBody VersionCommand versionCommand
    );


    String PATH_UPDATE_TASK = "/tasks/{taskId}";
    /**
     * PATCH /tasks/{taskId}
     *
     * @param taskId  (required)
     * @param updateTaskRequest  (required)
     * @return Task 수정 (status code 200)
     *         or 잘못된 요청 (status code 400)
     *         or 대상을 찾을 수 없음 (status code 404)
     *         or 버전 충돌 (status code 409)
     */
    @RequestMapping(
        method = RequestMethod.PATCH,
        value = TaskApi.PATH_UPDATE_TASK,
        produces = { "application/json", "application/problem+json" },
        consumes = { "application/json" }
    )
    ResponseEntity<TaskResponse> updateTask(
         @PathVariable("taskId") UUID taskId,
         @Valid @RequestBody UpdateTaskRequest updateTaskRequest
    );

}